+++ tests/TestCase.php
@@ -10,5 +10,6 @@ class TestCase extends \PHPUnit\Framework\TestCase
     public function setUp()
     {
         Wye::reset();
+        Wye::resetBacktraceForAllTests();
     }
 }