+++ CHANGELOG.md
@@ -5,8 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-
-- No pending changes
+### Added
+- Backtrace logging options to the Wye container
 
 ## [0.6.0] - 2018-01-28
 ### Added